Done - convert roadmap html elements to components
