+++ src/pages/roadmap/roadmap.js
@@ -3,21 +3,16 @@ import './roadmaps-module.css';
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
-import { MdDelete, MdEditDocument, MdAdd } from 'react-icons/md';
+import { MdAdd } from 'react-icons/md';
 
 import { getCurrentUserInformation } from '../../components/utils/userUtils';
-import EditRoadmap from './edit-roadmap/editRoadmap';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faClose } from '@fortawesome/free-solid-svg-icons';
+
 import PersonalRoadmapCard from '../../components/Roadmap-component/personal-roadmap-card-component/PersonalRoadmapCard';
-import EditRoadmapModal from './edit-roadmap/EditRoadmapModal';
 
 const Roadmap = () => {
     const [personalPersonalRoadmapList, setPersonalRoadmapList] = useState();
     const [isLoading, setIsLoading] = useState(true);
     const [triggerUpdate, setTriggerUpdate] = useState(true);
-    const [open, setOpen] = useState(false);
-    const [modalInformation, setModalInformation] = useState('');
 
     const user = getCurrentUserInformation();
     const userId = user.email;
@@ -40,21 +35,10 @@ const Roadmap = () => {
         fetchCurrentUserRoadmaps();
     }, [triggerUpdate]);
 
-    const handleOpen = (roadmapInformation) => {
-        setOpen(true);
-        setModalInformation(roadmapInformation);
-    };
-    const handleClose = () => {
-        setOpen(false);
-        setTriggerUpdate(!triggerUpdate)
-    };
-
     return (
         <div className="personal-roadmap-container">
             <div className="main-container">
                 <h1 className="personal-roadmap-title">My Road Maps</h1>
-
-
                 <p className="add-button">
                     <Link to="/roadmap/add" className="button-text">
                         New roadmap
@@ -70,17 +54,13 @@ const Roadmap = () => {
                         personalPersonalRoadmapList &&
                         personalPersonalRoadmapList.map((roadmapInformation) => {
                             return (
-                                
-                                <Link to={`/roadmap/${roadmapInformation.docId}`}>
-                                    <PersonalRoadmapCard
-                                        props={roadmapInformation}
-                                        handleClose={handleClose}
-                                        handleOpen={handleOpen}
-                                        stateChanger={setTriggerUpdate}
-                                        triggerUpdateProp={triggerUpdate}
-                                    />
-                                    {EditRoadmapModal}
-                                </Link>
+
+                                <PersonalRoadmapCard
+                                    props={roadmapInformation}
+                                    setTriggerUpdate={setTriggerUpdate}
+                                    triggerUpdate={triggerUpdate}
+                                />
+
                             );
                         })
                     )}
